Refactor into Spoon to enable easy cloning into Spoons folder
